Updated Changelog and the RedTribe Website.
v1.1 | Developing the website and adding backgrounds, images, and assets to make the HTML website stunning. Also creating the navigation bar to let people explore the website easily.Added the logo as an asset because, well, who is the website owned by or what is the company name/logo? This just confuses the person even more.