Background for the RedTribe Website
This helps from the current background we have is just straight red which messes with your eyes and hurts them. This will be a better substitute for the current background.

Also hurt my eyes when I looked at it. Immediately closed the tab and made this change right now. As the CEO myself, this will one hundred percent be a nice and comfortable substitute.