Deleted Background
We deleted the current background as the text would not go inside the image and failed to do so also. We tried all we could to add a nice and stunning background.

For everything to fit, we made a decision to delete the background from the image folder and the HTML code for a better look and design fit. We may keep it if we succeed, but for now this will be gone.

R.I.P. Background 2019-2019