r e d t r i b e      l o g o
We added the RedTribe Logo in the images folder, so you can take a look for yourself if you want to...for some weird reason. Anyways, this fits with our website and fits into everything.

Like the last update we made, we deleted the background and it was better off if everything fit inside the background, but we failed to do so. We made this a substitute so we can insure everything fits fine and our website isn't just a pile of mess.

We keep on striving to deliver the best experience possible on many devices, so we are just asking the HTML gods a simple and friendly please. We still look at the backgrounds grave. It had a quote that said "html gods said so." Pretty boring, right?

R.I.P. Background 2019-2019